harmonize

+++ samples/warp-dapr-rs/destroy.sh
@@ -23,4 +23,8 @@ kubectl delete secret servicebus-secret --ignore-not-found
 kubectl delete secret storage-secret --ignore-not-found
 kubectl delete -f ./dapr-components.yml
 
-helm uninstall receiver-standard-dapr receiver-express-dapr distributor-dapr
+if [ $PATTERN == 'shared' ]; then
+
+  helm uninstall receiver-standard-dapr receiver-express-dapr distributor-dapr
+
+fi
